Stop coercing JSON payload with .to_json
Since July the application has been saving JSON data as strings as opposed to actual JSON in the JSON column. This commit updates the ruby code to save the data in actual JSON format.

We also make it easier to setup development on a new machine by providing a convenient .env.template and more generic database.yml.

Finally we add the pry-byebug gem to the Gemfile to make debugging easier.